first progess moving state_types to better structure:Keep BaseAgentState and GraphConfig in shared_utils (foundational plumbing)      ☒ Move CogniAgentState and PlaywrightAgentState to shared_utils (generic, reusable)      ☒ Move ImageFlowState to src/cogni_image_gen/state_types.py (domain-specific)      ☒ Run uv run tox to verify all tests pass      ☒ Move system prompts to respective project directories (domain-specific)      ☒ Update imports in cogni_image_gen to use local state_types      ☒ Update imports in other projects to use shared_utils for generic states      ☒ Update test imports accordingly
